+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.2.2
+
+**Game Version**:
+- Added compatibility with V50 Patch 1
+
+**Bug Fixes**:
+- Fixed visual glitch where VR players would not appear to be sinking in mud
+- Fixed visual glitch where VR players who died in water got the underwater filter applied sporadically
+
+**Mod Compatibility**:
+- Fixed lighting culling issues when CullFactory is installed
+
 # 1.2.1
 
 ### V50 IS HERE